Fixed printout of script name
When running Test_contents_of_directory it prints out the names of each script. The point of this is to allow several scripts in the same language, that solve the same problem in different ways. While implementing this I discovered I had made a mistake in my regular expressions, which I fixed.